requirements.txt
pandas 
numpy 
sqlalchemy
scikit-learn
pyyaml
matplotlib
seaborn
statsmodels
imbalanced-learn
xgboost 
lightgbm
mlflow
